disposing is coming together

+++ packages/maestro_cli/lib/src/features/devices/devices_command.dart
@@ -7,6 +7,13 @@ import 'package:maestro_cli/src/common/common.dart';
 import 'package:maestro_cli/src/features/drive/device.dart';
 
 class DevicesCommand extends Command<int> {
+  DevicesCommand(DisposeScope parentDisposeScope)
+      : _disposeScope = DisposeScope() {
+    _disposeScope.disposed(parentDisposeScope);
+  }
+
+  final DisposeScope _disposeScope;
+
   @override
   String get name => 'devices';
 
@@ -15,10 +22,10 @@ class DevicesCommand extends Command<int> {
 
   @override
   Future<int> run() async {
-    final devices = await getDevices();
+    final devices = await _getDevices();
 
     if (devices.isEmpty) {
-      log.info('No devices attached');
+      log.warning('No devices attached');
       return 1;
     }
 
@@ -29,23 +36,29 @@ class DevicesCommand extends Command<int> {
     return 0;
   }
 
-  static Future<List<Device>> getDevices([DisposeScope? disposeScope]) async {
+  Future<List<Device>> _getDevices() async {
+    int? exitCode;
     final process = await Process.start(
       'flutter',
       ['devices', '--machine'],
       runInShell: true,
     );
 
-    disposeScope?.addDispose(() async {
+    _disposeScope.addDispose(() async {
+      if (exitCode != null) {
+        return;
+      }
+
       final msg = process.kill()
           ? 'Killed `flutter devices`'
           : 'Failed to kill `flutter devices`';
-      log.info(msg);
+      log.fine(msg);
     });
 
     var msg = '';
-    process.listenStdOut((line) => msg += line);
-    await process.exitCode;
+    process.listenStdOut((line) => msg += line).disposed(_disposeScope);
+    exitCode = await process.exitCode;
+    log.fine('exit code of `flutter devices`: $exitCode');
 
     final jsonOutput = jsonDecode(msg) as List<dynamic>;
 
@@ -70,4 +83,53 @@ class DevicesCommand extends Command<int> {
 
     return devices;
   }
+}
+
+@Deprecated('Cannot be disposed')
+Future<List<Device>> getDevices(DisposeScope disposeScope) async {
+  int? exitCode;
+  final process = await Process.start(
+    'flutter',
+    ['devices', '--machine'],
+    runInShell: true,
+  );
+
+  disposeScope.addDispose(() async {
+    if (exitCode != null) {
+      return;
+    }
+
+    final msg = process.kill()
+        ? 'Killed `flutter devices`'
+        : 'Failed to kill `flutter devices`';
+    log.fine(msg);
+  });
+
+  var msg = '';
+  process.listenStdOut((line) => msg += line).disposed(disposeScope);
+  exitCode = await process.exitCode;
+  log.fine('exit code of `flutter devices`: $exitCode');
+
+  final jsonOutput = jsonDecode(msg) as List<dynamic>;
+
+  final devices = <Device>[];
+  for (final deviceJson in jsonOutput) {
+    deviceJson as Map<String, dynamic>;
+
+    final targetPlatform = deviceJson['targetPlatform'] as String;
+    if (!targetPlatform.startsWith('android-') && targetPlatform != 'ios') {
+      continue;
+    }
+
+    devices.add(
+      Device(
+        name: deviceJson['name'] as String,
+        id: deviceJson['id'] as String,
+        targetPlatform: TargetPlatformX.fromString(targetPlatform),
+        real: !(deviceJson['emulator'] as bool),
+      ),
+    );
+  }
+
+  return devices;
 }
